real-vs-ai-detector/requirements.txt
# Core Framework
Flask==3.1.3
python-dotenv==1.2.2
psutil==7.2.2

# Deep Learning
tensorflow==2.21.0
keras==3.14.0
torch==2.11.0
torchvision==0.26.0
transformers==5.5.3
datasets==2.19.0
huggingface_hub==0.23.0

# Computer Vision
opencv-python==4.13.0.92
pillow==12.2.0
ultralytics==8.4.35

# ML & Scientific
numpy==2.4.4
scikit-learn==1.8.0
xgboost==3.2.0
matplotlib==3.10.8

# Similarity Search
faiss-cpu==1.13.2

# Inference Optimization
onnxruntime==1.24.4

# Production Server (Linux only)
# gunicorn==23.0.0
